Enable side get out on Wildcat & Puma
