+++ @twc_config_compatibility/addons/twc_gearFixes/config.cpp
@@ -1197,15 +1197,7 @@ class CfgVehicles {
 		};
 	};
 	
-	class Helicopter_Base_H;
-	class CUP_CH47F_base: Helicopter_Base_H
-	{
-		backRotorForceCoef = 0.9;
-		cyclicForwardForceCoef = 1.5;
-		cyclicAsideForceCoef = 1.3;
-		bodyFrictionCoef = 1.4;
-		armorStructural = 20;
-	};
+	#include "Helicopters.hpp"
 	
 //handling modifications
 	class Car;
